PythonBackend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
python-multipart
langchain
langchain-groq
dnspython
certifi
streamlit